AzureFunctions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
purviewcli
azure-storage-queue==2.1.0
rdflib
datacatalogtordf